use file_exists() instead of @-operator on unlink()

+++ redaxo/src/5.0/addons/phpmailer/lib/phpmailer.php
@@ -1237,12 +1237,16 @@ class PHPMailer {
         file_put_contents($file, $body); //TODO check this worked
         $signed = tempnam("", "signed");
         if (@openssl_pkcs7_sign($file, $signed, "file://".$this->sign_cert_file, array("file://".$this->sign_key_file, $this->sign_key_pass), NULL)) {
-          @unlink($file);
-          @unlink($signed);
+          if(file_exists($file))
+            unlink($file);
+          if(file_exists($signed))
+            unlink($signed);
           $body = file_get_contents($signed);
         } else {
-          @unlink($file);
-          @unlink($signed);
+          if(file_exists($file))
+            unlink($file);
+          if(file_exists($signed))
+            unlink($signed);
           throw new phpmailerException($this->Lang("signing").openssl_error_string());
         }
       } catch (phpmailerException $e) {
